refacto: name an anonymous class to allow serialization
